requirements.txt
flask==2.0.1
python-dotenv==0.19.0
supabase-py==0.0.2
requests==2.28.1
python-jose==3.3.0
httpx==0.23.0